backend/mcp/powerpoint_mcp/pyproject.toml
[project]
name = "powerpoint"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "loguru>=0.7.3",
    "mistralai>=1.9.7",
    "mcp>=1.0.0",
    "pillow>=11.3.0",
    "pydantic>=2.11.7",
    "python-dotenv>=1.1.1",
    "python-pptx>=1.0.2",
    "rich>=14.1.0",
    "typer>=0.16.1",
]

[project.scripts]
powerpoint-mcp = "server:main"
